LaCrosse removed unneeded start bit removal workaround
Moved time stamp to util.c
Alecto, fixed compile warnings (%%, time stamp, demod_arg init)
